refactor: add comments and rename a bit confusing functions
