+++ internal/doc.go
@@ -1,4 +1,4 @@
-// internal is internal packages for Ayd.
+// Package internal is internal packages for Ayd.
 //
 // Internal packages do not dependents on each other.
 // Dependencies to other package are implemented as a interface like scheme.Reporter or endpoint.Store.
