+++ internal/endpoint/doc.go
@@ -1,2 +1,2 @@
-// endpoint package implements the HTTP endpoints of Ayd.
+// Package endpoint implements the HTTP endpoints of Ayd such as status page.
 package endpoint
