+++ internal/scheme/doc.go
@@ -1,2 +1,4 @@
-// scheme package implements how to probe or send alert to the URL.
+// Package scheme implements how to probe or send alert to the URL.
+//
+// All scheme structs implement Prober and/or Alerter.
 package scheme
