+++ internal/scheme/scheme.go
@@ -1,5 +1,8 @@
 package scheme
 
+// SplitScheme splits scheme of URL.
+//
+// For example, "http-get" will splited to "http", '-', and "get".
 func SplitScheme(scheme string) (subScheme string, separator rune, variant string) {
 	for i, x := range scheme {
 		if x == '-' || x == '+' {
