+++ internal/store/doc.go
@@ -1,2 +1,2 @@
-// store package implements how to output the log, and how to restore from the log file.
+// Package store implements how to output the log, and how to restore from the log file.
 package store
