+++ internal/store/store_test.go
@@ -19,78 +19,6 @@ import (
 	api "github.com/macrat/ayd/lib-ayd"
 )
 
-func TestProbeHistoryMap(t *testing.T) {
-	m := make(store.ProbeHistoryMap)
-
-	for i := 1; i <= 100; i++ {
-		m.Append(&url.URL{Scheme: "dummy"}, api.Record{
-			CheckedAt: time.Now().Add(time.Duration(i) * time.Second),
-			Target:    &url.URL{Scheme: "dummy", Fragment: "append-test"},
-			Message:   fmt.Sprint(i),
-		})
-	}
-
-	if hs, ok := m["dummy:#append-test"]; !ok {
-		t.Errorf("failed to get history\n%#v", m)
-	} else if len(hs.Records) != store.PROBE_HISTORY_LEN {
-		t.Errorf("unexpected number of records: %d", len(hs.Records))
-	} else if hs.Records[len(hs.Records)-1].Message != "100" {
-		t.Errorf("unexpected message of latest record: %#v", hs.Records[len(hs.Records)-1])
-	}
-
-	for i := 1; i <= 10; i++ {
-		m.Append(&url.URL{Scheme: "dummy"}, api.Record{
-			CheckedAt: time.Now().Add(time.Duration(i) * time.Second),
-			Target:    &url.URL{Scheme: "dummy", Fragment: "append-test-another"},
-			Message:   fmt.Sprint(i),
-		})
-	}
-
-	if hs, ok := m["dummy:#append-test-another"]; !ok {
-		t.Errorf("failed to get history\n%#v", m)
-	} else if len(hs.Records) != 10 {
-		t.Errorf("unexpected number of records: %d", len(hs.Records))
-	} else if hs.Records[len(hs.Records)-1].Message != "10" {
-		t.Errorf("unexpected message of latest record: %#v", hs.Records[len(hs.Records)-1])
-	}
-
-	for i := 1; i <= 10; i++ {
-		m.Append(&url.URL{Scheme: "dummy"}, api.Record{
-			CheckedAt: time.Now().Add(time.Duration(-i) * time.Second),
-			Target:    &url.URL{Scheme: "dummy", Fragment: "append-test-reverse"},
-			Message:   fmt.Sprint(i),
-		})
-	}
-
-	if hs, ok := m["dummy:#append-test-reverse"]; !ok {
-		t.Errorf("failed to get history\n%#v", m)
-	} else if len(hs.Records) != 10 {
-		t.Errorf("unexpected number of records: %d", len(hs.Records))
-	} else if hs.Records[len(hs.Records)-1].Message != "1" {
-		t.Errorf("unexpected message of latest record: %#v", hs.Records[len(hs.Records)-1])
-	}
-
-	timestamp := time.Now()
-	m.Append(&url.URL{Scheme: "dummy"}, api.Record{
-		CheckedAt: timestamp,
-		Target:    &url.URL{Scheme: "dummy", Fragment: "append-test-same-time"},
-		Message:   "first",
-	})
-	m.Append(&url.URL{Scheme: "dummy"}, api.Record{
-		CheckedAt: timestamp,
-		Target:    &url.URL{Scheme: "dummy", Fragment: "append-test-same-time"},
-		Message:   "second",
-	})
-
-	if hs, ok := m["dummy:#append-test-same-time"]; !ok {
-		t.Errorf("failed to get history\n%#v", m)
-	} else if len(hs.Records) != 2 {
-		t.Errorf("unexpected number of records: %d", len(hs.Records))
-	} else if hs.Records[len(hs.Records)-1].Message != "second" {
-		t.Errorf("unexpected message of latest record: %#v", hs.Records[len(hs.Records)-1])
-	}
-}
-
 type Buffer struct {
 	sync.Mutex
