+++ internal/testutil/doc.go
@@ -1,3 +1,3 @@
-// testutil package implements utilities for testing.
+// Package testutil implements utilities for testing.
 // This package is not used by production code.
 package testutil
